requirements.txt
absl-py==1.4.0
anyascii==0.3.1
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
blis==0.7.9
bokeh==3.1.0
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
confection==0.0.4
contourpy==1.0.7
contractions==0.1.73
cycler==0.11.0
cymem==2.0.7
Cython==0.29.32
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
flatbuffers==23.1.21
fonttools==4.38.0
fst-pso==1.8.1
FuzzyTM==2.0.5
gast==0.4.0
gensim==4.3.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
gunicorn==20.1.0
h5py==3.8.0
idna==3.4
imageio==2.26.0
ipykernel==6.21.2
ipython==8.10.0
ipywidgets==8.0.4
itsdangerous==2.1.2
jax==0.4.4
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyterlab-widgets==3.0.5
keras==2.12.0rc1
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.1
libclang==15.0.6.1
lime
livelossplot==0.5.5
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
miniful==0.0.6
murmurhash==1.0.9
nest-asyncio==1.5.6
networkx==3.0
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathy==0.10.1
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
plotly==5.13.1
preshed==3.0.8
prompt-toolkit==3.0.37
protobuf==4.22.0
psutil==5.9.4
pure-eval==0.2.2
pyahocorasick==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.5
pyFUME==0.2.25
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.0
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.12.2
simpful==2.10.0
six==1.16.0
smart-open==6.3.0
srsly==2.4.5
stack-data==0.6.2
streamlit
tenacity==8.2.2
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0rc0
tensorflow-estimator==2.12.0rc0
tensorflow-intel==2.12.0rc0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
textsearch==0.0.24
thinc==8.1.7
threadpoolctl==3.1.0
tifffile==2023.2.28
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.14
wasabi==1.1.1
wcwidth==0.2.6
Werkzeug==2.2.3
widgetsnbextension==4.0.5
wrapt==1.15.0
xyzservices==2023.2.0
